# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Semillero",
        "description": "API para el sistema de Semillero",
        "contact": {
            "email": "tu_email@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/semilleros": {
            "get": {
                "tags": [
                    "CRUD Semilleros"
                ],
                "summary": "Listar todos los semilleros",
                "operationId": "e82b9d0ae33e4ea263875e498b0940ea",
                "responses": {
                    "200": {
                        "description": "Listado de semilleros"
                    }
                }
            },
            "post": {
                "tags": [
                    "CRUD Semilleros"
                ],
                "summary": "Crear un nuevo semillero",
                "operationId": "4d9fd225c6ab0a9c432e3c84de3b91dd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "descripcion",
                                    "programa_id",
                                    "profesor_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "programa_id": {
                                        "type": "integer"
                                    },
                                    "profesor_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Semillero creado correctamente"
                    }
                }
            }
        },
        "/api/semilleros/{id}": {
            "get": {
                "tags": [
                    "CRUD Semilleros"
                ],
                "summary": "Consultar un semillero por ID",
                "operationId": "5af90950553e1ed1d43222cdc122ee5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del semillero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Semillero encontrado"
                    },
                    "404": {
                        "description": "Semillero no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "CRUD Semilleros"
                ],
                "summary": "Actualizar un semillero por ID",
                "operationId": "a27a74b1df9672b74f328bc94570d50c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del semillero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "descripcion",
                                    "programa_id",
                                    "profesor_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "programa_id": {
                                        "type": "integer"
                                    },
                                    "profesor_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Semillero actualizado correctamente"
                    },
                    "404": {
                        "description": "Semillero no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "CRUD Semilleros"
                ],
                "summary": "Eliminar un semillero por ID",
                "operationId": "65ba56f23809aef3cac603d90fa8595e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del semillero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Semillero eliminado correctamente"
                    },
                    "404": {
                        "description": "Semillero no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}